requirements.txt
Django==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
gunicorn==21.2.0
